Fix the colour coding/CSS styling for document types
Since we've refactored the document definitions completely into objects,
their names are now taken from the names of the constructor functions
and can't be overridden. This means that we now have to either use
these constructor names in our CSS or else define a new property called
className or something.

In the interests of convention over configuration, I've opted for the
former.