Use the "proper" way of creating tooltip.
According to the SVG specification, tooltips are supposed to be created using
a <title> element. Unfortunately this doesn't seem to work for reasons that
completely befuddle me, because I've come across jsFiddle samples that work
just fine here. The <title> attribute works fine in Firefox and Chrome though.

For the time being, add both a <title> element and a title=... attribute.
We'll sort this one out later.